moved filter to filter package
